Fix qgRound() 1st attempt
see http://osgeo-org.1560.x6.nabble.com/QGIS-Developer-On-precision-in-qgsRound-3-3-0-master-td5377378.html
see https://github.com/qgis/QGIS/pull/7836#issuecomment-419718554

+++ src/core/qgis.h
@@ -287,10 +287,10 @@ inline bool qgsDoubleNearSig( double a, double b, int significantDigits = 10 )
  *
  * \since QGIS 3.0
  */
-inline double qgsRound( double number, double places )
+inline double qgsRound( double number, int places )
 {
-  int scaleFactor = std::pow( 10, places );
-  return static_cast<double>( static_cast<qlonglong>( number * scaleFactor + 0.5 ) ) / scaleFactor;
+  qlonglong scaleFactor = std::pow( 10, places );
+  return static_cast<double>( static_cast<qulonglong>( number * scaleFactor + 0.5 ) ) / scaleFactor;
 }
